Reproducible TFile: Sort StreamerInfos
The StreamerInfo list is dependend on the order in which classes were
loaded so it can differ depending on the code executed before writing a
file.

We can mitigate this by sorting the TStreamerInfo by class name.
TStreamerInfo doesn't seem to have a Compare() and the class name might
not be the most general comparison order so lets sort them in a
temporary vector.